Add files via upload
The information provided here is being provided as a courtesy to Roadster enthusiasts and was created during the design phase of the Roadster for research and development. It isn't manufacturer reference or repair and maintenance material, and may not accurately reflect the actual production models or parts sold. If you use this information, it's your responsibility to ensure that you follow all laws and safety protocols as we don't provide any warranties on any work done by non-Tesla personnel whether or not you use the information provided here. You also understand that if you make or design parts or create new repairs or procedures based on this information, we won't be responsible for them and they won't be genuine Tesla parts or accessories or Tesla-approved procedures.